@@ -1,17 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
-    <PackageReference Include="TauCode.Algorithms.Graphs" Version="1.7.0" />
-    <PackageReference Include="TauCode.Data" Version="1.7.0" />
-    <PackageReference Include="TauCode.Extensions" Version="1.7.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="TauCode.Algorithms.Graphs" Version="2.0.1" />
+    <PackageReference Include="TauCode.Extensions" Version="2.0.1" />
   </ItemGroup>
 
 </Project>
